stand: first version
This implements a rocking stand to etch PCBs. It is a C reimplementation
of my earlier 12f675-based one, but with an AVR this time. 3 potentiometers
are used :
  - p1 = center
  - p2 = speed
  - p3 = amplitude

This may also be used to test servos and their speed to move in any direction.
The code is roughly 500 bytes and works equally well in an ATtiny13A or an
ATtiny85. It is important to note that the timer calibration loop takes 5
cycles per iteration with gcc 5.4 and 4 with gcc 4.3, so building it with
gcc 4.3 will not correctly work and will require to change a ratio in the
us2loops() function.