stand: increase range for the centering
In order to be able to more easily reuse servos with broken gears, we
need to cover a larger range for the center. This also requires to
increase the possible amplitude around the 1500 us center, and in order
to limit measurement noise, we'll retrieve the full 10-bit ADC values
for the center and the speed instead of using only 8-bit shifted left,
which is prone to noise. Now it is possible to make the servo oscillate
on any area.